+++ lazarkit/src/main/java/lol/lazar/lazarkit/panels/configurables/ArrayElement.kt
@@ -1,6 +1,6 @@
 package lol.lazar.lazarkit.panels.configurables
 
-import lol.lazar.lazarkit.panels.data.GenericTypeJson
+import lol.lazar.lazarkit.panels.json.GenericTypeJson
 import java.lang.reflect.Array
 
 class ArrayElement(
