+++ lazarkit/src/main/java/lol/lazar/lazarkit/panels/configurables/GenericField.kt
@@ -1,7 +1,7 @@
 package lol.lazar.lazarkit.panels.configurables
 
 import lol.lazar.lazarkit.panels.configurables.annotations.IgnoreConfigurable
-import lol.lazar.lazarkit.panels.data.GenericTypeJson
+import lol.lazar.lazarkit.panels.json.GenericTypeJson
 import java.lang.reflect.Array
 import java.lang.reflect.Field
 import java.util.UUID
