+++ lazarkit/src/main/java/lol/lazar/lazarkit/panels/json/OpModeJsonData.kt
@@ -1,4 +1,4 @@
-package lol.lazar.lazarkit.panels.data
+package lol.lazar.lazarkit.panels.json
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
