+++ lazarkit/src/main/java/lol/lazar/lazarkit/panels/server/Socket.kt
@@ -4,26 +4,26 @@ import com.qualcomm.hardware.lynx.LynxModule
 import fi.iki.elonen.NanoWSD
 import kotlinx.serialization.PolymorphicSerializer
 import lol.lazar.lazarkit.panels.GlobalData
-import lol.lazar.lazarkit.panels.OpModeData
+import lol.lazar.lazarkit.panels.integration.OpModeData
 import lol.lazar.lazarkit.panels.configurables.Configurables
-import lol.lazar.lazarkit.panels.data.ActiveOpMode
-import lol.lazar.lazarkit.panels.data.BatteryVoltage
-import lol.lazar.lazarkit.panels.data.Canvas
-import lol.lazar.lazarkit.panels.data.GetActiveOpModeRequest
-import lol.lazar.lazarkit.panels.data.GetJvmFieldsRequest
-import lol.lazar.lazarkit.panels.data.GetOpModesRequest
-import lol.lazar.lazarkit.panels.data.InitOpModeRequest
-import lol.lazar.lazarkit.panels.data.JSONData
-import lol.lazar.lazarkit.panels.data.ReceivedJvmFields
-import lol.lazar.lazarkit.panels.data.ReceivedOpModes
-import lol.lazar.lazarkit.panels.data.StartActiveOpModeRequest
-import lol.lazar.lazarkit.panels.data.StopActiveOpModeRequest
-import lol.lazar.lazarkit.panels.data.TelemetryPacket
-import lol.lazar.lazarkit.panels.data.TestObject
-import lol.lazar.lazarkit.panels.data.TimeObject
-import lol.lazar.lazarkit.panels.data.UpdatedJvmFields
-import lol.lazar.lazarkit.panels.data.json
-import lol.lazar.lazarkit.panels.data.toJson
+import lol.lazar.lazarkit.panels.json.ActiveOpMode
+import lol.lazar.lazarkit.panels.json.BatteryVoltage
+import lol.lazar.lazarkit.panels.json.Canvas
+import lol.lazar.lazarkit.panels.json.GetActiveOpModeRequest
+import lol.lazar.lazarkit.panels.json.GetJvmFieldsRequest
+import lol.lazar.lazarkit.panels.json.GetOpModesRequest
+import lol.lazar.lazarkit.panels.json.InitOpModeRequest
+import lol.lazar.lazarkit.panels.json.JSONData
+import lol.lazar.lazarkit.panels.json.ReceivedJvmFields
+import lol.lazar.lazarkit.panels.json.ReceivedOpModes
+import lol.lazar.lazarkit.panels.json.StartActiveOpModeRequest
+import lol.lazar.lazarkit.panels.json.StopActiveOpModeRequest
+import lol.lazar.lazarkit.panels.json.TelemetryPacket
+import lol.lazar.lazarkit.panels.json.TestObject
+import lol.lazar.lazarkit.panels.json.TimeObject
+import lol.lazar.lazarkit.panels.json.UpdatedJvmFields
+import lol.lazar.lazarkit.panels.json.json
+import lol.lazar.lazarkit.panels.json.toJson
 import org.firstinspires.ftc.robotcore.external.navigation.VoltageUnit
 import java.io.IOException
 import java.text.SimpleDateFormat
