+++ lazarkit/src/test/java/lol/lazar/lazarkit/panels/json/JSONDataTest.kt
@@ -1,4 +1,4 @@
-package lol.lazar.lazarkit.panels.data
+package lol.lazar.lazarkit.panels.json
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
